Extract splinechar.h
